pyproject.toml
[project]
name = "pymedx"
version = "0.5.0"  # semantic-release
description = "PyMedX is a tool set for handling multimedia files."
readme = "docs/index.md"
authors = [
  { name = "Gijs Wobben", email = "gijswobben@gmail.com" },
  { name = "Ivan Ogasawara", email = "ivan.ogasawara@gmail.com" },
  { name = "Ever Vino", email = "evervino00@gmail.com" },
]
packages = [
  {include = "pymedx", from="src"},
]
license = "MIT"
exclude = [
  ".git/*",
  ".env*",
]
requires-python = ">3.9,<4"
dependencies = [
  "requests (>=2.20.0)",
  "lxml (>=5.1.0)",
]

[tool.poetry.group.dev.dependencies]
python = ">3.9,<4"
python-dotenv = ">=1.0"
pytest = ">=7.3.2"
pytest-cov = ">=4.1.0"
coverage = ">=7.2.7"
pre-commit = ">=3.3.2"
ruff = ">=0.9.5"
mypy = ">=1.6.0"
bandit = ">=1.7.5"
vulture = ">=2.9.1"
mccabe = ">=0.6.1"
ipykernel = ">=6.0.0"
ipython = ">=7.0.0"
mkdocs = ">=1.4.3"
mkdocs-exclude = ">=1.0.2"
mkdocs-jupyter = ">=0.24.1"
mkdocs-literate-nav = ">=0.6.0"
mkdocs-macros-plugin = ">=0.7.0,<1"
mkdocs-material = ">=9.1.15"
mkdocstrings = ">=0.21.2"
mkdocstrings-python = ">=1.1.2"
jupyterlab = ">=4.0.5"
makim = "1.12.0"
requests-cache = ">=1"
types-lxml = "^2023.10.21"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.vulture]
exclude = ["tests"]
ignore_decorators = []
ignore_names = []
make_whitelist = true
min_confidence = 80
paths = ["src/pymedx"]
sort_by_size = true
verbose = false

[tool.ruff]
line-length = 79
force-exclude = true
src = ["./"]
ignore = ["PLR0913"]
exclude = [
  'docs',
  'examples',
]
select = [
  "E",   # pycodestyle
  "F",   # pyflakes
  "D",   # pydocstyle
  "YTT", # flake8-2020
  "PL",  # PL
  "RUF", # Ruff-specific rules
  "I001", # isort
]
fix = true

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.ruff.isort]
# Use a single line between direct and from import
lines-between-types = 1

[tool.ruff.format]
quote-style = "double"


[tool.mypy]
no_strict_optional = false
exclude = [
    '^docs/$',
    '^examples/$',
]

[[tool.mypy.overrides]]
module = [
  "requests",
]
ignore_missing_imports = true
